requirements.txt
aiogram==3.17.0
spotipy==2.23.0
python-dotenv==1.0.1
aiohttp==3.9.3
ytmusicapi==1.4.2
thefuzz==0.22.1
python-levenshtein==0.25.0
mutagen==1.47.0